requirements.txt
Django==5.1.5
djangorestframework==3.15.2
pandas==2.2.3
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
black==24.10.0
isort==5.13.2
flake8==7.1.1
